refactor de UserController UserModel
